♻️ refactor: common 디렉토리 분리, 상수 파일 따로 분리
