+++ rss-notifier/src/common/constant.ts
@@ -1,0 +1,6 @@
+export const CONNECTION_LIMIT = 50;
+export const INSERT_ID = "insertId";
+export const redisConstant = {
+  FEED_RECENT_ALL_KEY: "feed:recent:*",
+  FEED_RECENT_KEY: "feed:recent",
+};
