+++ rss-notifier/src/common/types.ts
