+++ rss-notifier/src/rss-notifier.ts
@@ -1,12 +1,12 @@
-import logger from "./logger.js";
+import logger from "./common/logger.js";
 import "dotenv/config";
 import {
   selectAllRss,
   insertFeeds,
   deleteRecentFeedStartId,
   setRecentFeedList,
-} from "./db-access.js";
-import { RssObj, FeedDetail, RawFeed } from "./types.js";
+} from "./common/db-access.js";
+import { RssObj, FeedDetail, RawFeed } from "./common/types.js";
 import { XMLParser } from "fast-xml-parser";
 import { parse } from "node-html-parser";
 import { unescape } from "html-escaper";
